fix(rating-bar): Adjust vertical alignment and switch to flex
## Description:
This commit modifies the rating-bar to use flex as we can switch to
'column' direction in a media query and achieve the mobile appearance
with less changes.
Switching to flex also has the added benefit of allowing us to
vertically center the text in a more effective manner.
Additionally, it adjusts the padding to make the rating-bar
appear taller

Signed-off-by: Aditya Rao <adirlzr@gmail.com>